requirements.txt
# Core Flask dependencies
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0

# Browser automation for Claude subscription auth
selenium==4.15.0
undetected-chromedriver==3.5.0
playwright==1.40.0

# Claude interaction (no API key needed)
requests==2.31.0
beautifulsoup4==4.12.2
websocket-client==1.6.4

# Database
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23

# Authentication
flask-login==0.6.3
python-jose==3.3.0
python-dotenv==1.0.0

# WebSocket support
python-engineio==4.8.0
eventlet==0.33.3

# File handling
werkzeug==3.0.1
python-magic==0.4.27

# Markdown and document processing
markdown==3.5.1
pymdown-extensions==10.5
beautifulsoup4==4.12.2

# JSON and API
requests==2.31.0
jsonschema==4.20.0

# Development tools
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0

# Production server
gunicorn==21.2.0
gevent==23.9.1

# Token estimation
tiktoken>=0.11.0